execute: allow detached mode ('command &')
A process  can be started and detached from shell. This way the process is running but the shell does not get blocked.

Example:
- `npm start` starts a node process and wait until it is finished
- `npm start &` starts the process but the shell can execute other commands after that

In node, the `exec('npm start &')` does not execute the '&'. So the promise version waits until ended.

To allow using execute() for normal and detached commands, the workaruind is to use unref() and return inmediately instead waiting for resolving the promise.